requirements.txt
fake-bpy-module-latest
lib3mf